Rename `helpers/{view-filter => filter-annotations}`
Rename the module to match the main function it exports. This hopefully make it
more obvious what the module does.
